Secure Login
This is the signup/sign-in page of my final application, which uses Firebase to store credentials, ensuring secure login. During the sign-in process, it sends an email verification, and until this is confirmed, it's not possible to log in. Additionally, there is a 'Forgot Password' page where users can enter their email to receive a reset link for confirmation.the interface is shown